Statements keep references to their database
Since statements must all be finalized before a database closes, the statements should keep a reference to the database object. The database object cannot be garbage collected until the statements are (or until they're all finalized, but that only happens during garbage collection for statements at the moment) since it raises an error during program shutdown and may even possibly corrupt the database, or leave open journals. With statements keeping a reference to the database object, it ensures they will be garbage collected before the database object itself.

It also allows for having an interface that only exposes prepared statements, and not the database object itself, if you really don't like carting the database object around manually or sticking it in a global variable.