Respect CC environment variable in Makefile
The GNU C compiler does not by default come with every platform.
Hardcoding it makes the library extremely unportable. This checks
whether CC is declared in the environment. If so, it uses that. In any
other case it defaults to /usr/bin/cc which, on most Linux distros
defaults to GCC. On FreeBSD, which was the issue, it defaults to
clang.