Merge `Statement.bind` overloads together
Using template constraints works well when the constraints
describe mutually-exclusive set of types.
However, template constraints do not scale when some types can
match multiple overloads, such as when check for the existence
of a method. When constraints are not mutually exclusive,
one need to describe the priority of such constraints,
which is more easily done in a single function.